+++ oalg-abg/src/OAlg/AbelianGroup/KernelsAndCokernels.hs
@@ -53,6 +53,7 @@ import OAlg.Structure.Number
 
 import OAlg.Adjunction
 
+import OAlg.Limes.Universal
 import OAlg.Limes.Cone
 import OAlg.Limes.Definition
 import OAlg.Limes.Limits