+++ oalg-base/src/OAlg/Entity/Slice/Definition.hs
@@ -27,6 +27,9 @@ module OAlg.Entity.Slice.Definition
     -- * Slice
     Slice(..), slice, slSiteType
 
+    -- ** Duality
+  , coSlice, coSliceInv
+
     -- * Factor
   , SliceFactor(..), slfFactor, slfIndex
   
@@ -48,13 +51,6 @@ module OAlg.Entity.Slice.Definition
     -- ** Injective
   , slfLimitsInjective
 
-    -- * Liftables
-  , Liftable(..), liftBase, lift
-  , LiftableException(..)
-
-    -- ** Duality
-  , coLiftable
-  
     -- * X
   , xSliceTo, xSliceFrom
   , xosXOrtSiteToSliceFactorTo
@@ -449,7 +445,7 @@ dgSlfToSlicePoint d = ConeInjective d' t cs where
 
 -- | the induced 'Injective' limes for 'SliceFactor'. 
 slfLimesInjective :: (Multiplicative c, Sliced i c)
-  => Limits Mlt Injective t n m c
+  => Limits Limes Mlt Injective t n m c
   -> Diagram t n m (SliceFactor To i c)
   -> Limes Mlt Injective t n m (SliceFactor To i c)
 slfLimesInjective l dgSlf = LimesInjective slfLim slfUniv where
@@ -467,7 +463,7 @@ slfLimesInjective l dgSlf = LimesInjective slfLim slfUniv where
 
 -- | the induced 'Injective' 'Limits'.
 slfLimitsInjective :: (Multiplicative c, Sliced i c)
-  => Limits Mlt Injective t n m c -> Limits Mlt Injective t n m (SliceFactor To i c)
+  => Limits Limes Mlt Injective t n m c -> Limits Limes Mlt Injective t n m (SliceFactor To i c)
 slfLimitsInjective lms = Limits $ slfLimesInjective lms
 
 --------------------------------------------------------------------------------
@@ -591,110 +587,6 @@ xosAdjTerminal w xos@(XEnd xp xf) = XEnd xp' xf where
   i   = slfIndex xos
   s   = SliceTo i (one $ slicePoint i)
 
-
---------------------------------------------------------------------------------
--- LiftableException -
-
--- | liftable exceptions which are sub exceptions of 'SomeOAlgException'.
-data LiftableException
-  = NotLiftable
-  deriving (Eq,Show)
-
-instance Exception LiftableException where
-  toException   = oalgExceptionToException
-  fromException = oalgExceptionFromException
-
---------------------------------------------------------------------------------
--- Liftable -
-
--- | liftable slices.
---
--- __Property__ Let @l@ be in @'Liftable' __p__ __i__ __c__@ for an @__i__@-sliced 'Oriented'
--- structure @__c__@, then holds:
---
--- (1) If @l@ matches @'LiftableFrom' c lift@, then holds:
--- For all @f@ in @'Slice' 'From' __i__ __c__@ holds:
---
---     (1) If @'end' c '/=' 'end' ('slice' f)@ then the evaluation of @lift f@ ends up in a
---     'NotLiftable'-exception.
---
---     (2) If @'end' c '==' 'end' ('slice' f)@ then @lift f@ is 'valid' and
---     @'slice' f '==' c '*' 'slice' (lift f)@.
---
--- (2) If @l@ matches @'LiftableTo' c lift@, then holds:
--- For all @t@ in @'Slice' 'To' __i__ __c__@ holds:
---
---     (1) If @'start' c '/=' 'start' ('slice' t)@ then the evaluation of @lift t@ ends up in a
---     'NotLiftable'-exception.
---
---     (2) If @'start' c '==' 'start' ('slice' t)@ then @lift t@ is 'valid' and
---     @'slice' t '==' 'slice' (lift l) '*' c@.
-data Liftable s i c where
-  LiftableFrom :: Sliced i c => c -> (Slice From i c -> Slice From i c) -> Liftable From i c
-  LiftableTo :: Sliced i c => c -> (Slice To i c -> Slice To i c) -> Liftable To i c
-
-instance Show c => Show (Liftable s i c) where
-  show (LiftableFrom c _) = join ["LiftableFrom (",show c,") lift"]
-  show (LiftableTo c _)   = join ["LiftableTo (",show c,") lift"]
-
---------------------------------------------------------------------------------
--- Liftable - Dual -
-
-type instance Dual (Liftable s i c) = Liftable (Dual s) i (Op c)
-
-coLiftable :: Dual (Dual s) :~: s -> Liftable s i c -> Dual (Liftable s i c)
-coLiftable r (LiftableFrom c lift) = LiftableTo (Op c) (coSlice . lift . coSliceInv r)
-coLiftable r (LiftableTo c lift) = LiftableFrom (Op c) (coSlice . lift . coSliceInv r)
-
---------------------------------------------------------------------------------
--- Liftable - Valid -
-
-relLiftableFrom :: Multiplicative c => i c -> XOrtOrientation c -> Liftable From i c -> Statement
-relLiftableFrom i xo (LiftableFrom c lift)
-  = And [ Label "c" :<=>: valid c
-        , Forall xf (\f
-            -> And [ Label "f" :<=>: valid f
-                   , let f' = lift f in case end c == end (slice f) of
-                       False -> (valid f' :=> throw implError)
-                                  `Catch` (\e -> case e of NotLiftable -> SValid)
-                       True  -> (slice f == c * slice f')
-                                  :?> Params ["c":=show c,"f":=show f,"lift f":=show f']
-                   ]
-                    )
-        ]
-    
-  where implError = ImplementationError "unliftable dos not throw a NotLiftable-exception"
-        ip = slicePoint i
-  
-        xf = amap1 (SliceFrom i)
-           $ xOneOfXW [ (9,xoArrow xo (ip :> end c))
-                      , (1,xoPoint xo >>= xoArrow xo . (ip:>))
-                      ]
-
-relLiftable :: Multiplicative c => XOrtOrientation c -> Liftable s i c -> Statement
-relLiftable xo l = case l of
-  LiftableFrom _ _ -> relLiftableFrom unit1 xo l
-  LiftableTo _ _   -> relLiftable (coXOrtOrientation xo) (coLiftable Refl l)
-  
-instance (Multiplicative c, XStandardOrtOrientation c)
-  => Validable (Liftable s i c) where
-  valid l = Label "Liftable" :<=>: relLiftable xStandardOrtOrientation l
-                                      
---------------------------------------------------------------------------------
--- liftBase -
-
--- | the underlying factor.
-liftBase :: Liftable s i c -> c
-liftBase (LiftableFrom c _) = c
-liftBase (LiftableTo c _) = c
-
---------------------------------------------------------------------------------
--- lift -
-
--- | the lifting map.
-lift :: Liftable s i c -> Slice s i c -> Slice s i c
-lift (LiftableFrom _ l) = l
-lift (LiftableTo _ l) = l
 
 --------------------------------------------------------------------------------
 -- Slice - Structure -