+++ oalg-base/CHANGELOG.md
@@ -145,3 +145,12 @@ Empty2 to EntEmpty2, empty2 to fromEmpty2.
 - Elimination of the idioms 'Embeddable' and 'Forgetful' such that only the Idiom
 'Transformable' and 'Transformable...' are needed.
 - Constraints for Transformation being Vectorial.
+
+## [2.3.0.0] - 2025-02-14
+
+Indroducing generic limits.
+
+### Added
+- The class Uiversal and UniversalApplicative in the module OAlg.Limes.Universal for generic limits.
+
+